@@ -1,26 +1,24 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk.Web">
 
 	<PropertyGroup>
-		<TargetFramework>net7.0</TargetFramework>
+		<TargetFramework>net8.0</TargetFramework>
 		<Nullable>disable</Nullable>
 		<ImplicitUsings>enable</ImplicitUsings>
 	</PropertyGroup>
 
 	<ItemGroup>
 		<PackageReference Include="AutoMapper" Version="13.0.1" />
-		<PackageReference Include="FluentValidation" Version="11.9.0" />
-		<PackageReference Include="Microsoft.EntityFrameworkCore.Design" Version="7.0.12">
+		<PackageReference Include="FluentValidation" Version="11.9.2" />
+		<PackageReference Include="Microsoft.AspNetCore.Authentication.JwtBearer" Version="8.0.6" />
+		<PackageReference Include="Microsoft.EntityFrameworkCore.Design" Version="8.0.6">
 			<PrivateAssets>all</PrivateAssets>
 			<IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
 		</PackageReference>
-		<PackageReference Include="Microsoft.VisualStudio.Web.CodeGeneration.Design" Version="7.0.12" />
+		<PackageReference Include="Microsoft.VisualStudio.Web.CodeGeneration.Design" Version="8.0.2" />
 	</ItemGroup>
 
 	<ItemGroup>
 	  <ProjectReference Include="..\..\..\building blocks\EMS.Core\EMS.Core.csproj" />
-	  <ProjectReference Include="..\EMS.WebApp.Business\EMS.WebApp.Business.csproj" />
-	  <ProjectReference Include="..\EMS.WebApp.Data\EMS.WebApp.Data.csproj" />
-	  <ProjectReference Include="..\EMS.WebApp.Identity\EMS.WebApp.Identity.csproj" />
 	</ItemGroup>
 
 </Project>
