@@ -1,6 +1,5 @@
 ﻿using EMS.Core.Handlers;
 using EMS.WebApi.Business.Handlers;
-using EMS.WebApi.Business.Utils;
 using Microsoft.Extensions.DependencyInjection;
 
 namespace EMS.WebApi.Business.Configuration;
@@ -9,10 +8,12 @@ public static class DependencyInjectionConfig
 {
     public static IServiceCollection RegisterBusinessServices(this IServiceCollection services)
     {
-        services.AddScoped<IAspNetUser, AspNetUser>();
-
+        services.AddScoped<ICompanyHandler, CompanyHandler>();
+        services.AddScoped<IPlanHandler, PlanHandler>();
         services.AddScoped<IEmployeeHandler, EmployeeHandler>();
         services.AddScoped<IClientHandler, ClientHandler>();
+        services.AddScoped<IServiceHandler, ServiceHandler>();
+        services.AddScoped<IServiceAppointmentHandler, ServiceAppointmentHandler>();
         services.AddScoped<IProductHandler, ProductHandler>();
 
         return services;
