@@ -3,14 +3,13 @@ using EMS.Core.Notifications;
 using EMS.Core.Requests.Products;
 using EMS.Core.Responses;
 using EMS.Core.Responses.Products;
+using EMS.Core.User;
 using EMS.WebApi.Business.Interfaces.Repositories;
 using EMS.WebApi.Business.Mappings;
-using EMS.WebApi.Business.Services;
-using EMS.WebApi.Business.Utils;
 
 namespace EMS.WebApi.Business.Handlers;
 
-public class ProductHandler : MainService, IProductHandler
+public class ProductHandler : BaseHandler, IProductHandler
 {
     public readonly IProductRepository _productRepository;
     public readonly ICompanyRepository _companyRepository;
