@@ -1,174 +1,170 @@
-﻿using EMS.Core.Handlers;
-using EMS.Core.Requests.Products;
-using EMS.Core.Responses;
-using EMS.Core.Responses.Products;
-using EMS.WebApp.Business.Notifications;
-using EMS.WebApp.Business.Utils;
-using EMS.WebApp.MVC.Models;
+﻿using EMS.Core.Notifications;
 using Microsoft.AspNetCore.Authorization;
-using Microsoft.AspNetCore.Mvc;
 
 namespace EMS.WebApp.MVC.Controllers;
 
 [Authorize]
 public class ProductsController : MainController
 {
-    private readonly IProductHandler _productHandler;
-    private readonly IAspNetUser _appUser;
+    //private readonly IProductHandler _productHandler;
+    //private readonly IAspNetUser _appUser;
 
-    public ProductsController(INotifier notifier, IProductHandler productHandler, IAspNetUser appUser) : base(notifier)
+    public ProductsController(INotifier notifier) : base(notifier)
     {
-        _productHandler = productHandler;
-        _appUser = appUser;
     }
+    //public ProductsController(INotifier notifier, IProductHandler productHandler, IAspNetUser appUser) : base(notifier)
+    //{
+    //    _productHandler = productHandler;
+    //    _appUser = appUser;
+    //}
 
-    public async Task<IActionResult> Index([FromQuery] int page = 1, [FromQuery] string q = null)
-    {
-        var ps = 10;
-        var request = new GetAllProductsRequest { PageNumber = page, PageSize = ps, Query = q };
-        var response = await _productHandler.GetAllAsync(request);
+    //public async Task<IActionResult> Index([FromQuery] int page = 1, [FromQuery] string q = null)
+    //{
+    //    var ps = 10;
+    //    var request = new GetAllProductsRequest { PageNumber = page, PageSize = ps, Query = q };
+    //    var response = await _productHandler.GetAllAsync(request);
 
-        //var mappedProducts = new PagedViewModel<ProductResponse>
-        //{
-        //    List = response.Data,
-        //    PageIndex = request.PageNumber,
-        //    PageSize = request.PageSize,
-        //    Query = request.Query,
-        //    TotalResults = response.TotalCount
-        //};
-        //ViewBag.Search = q;
-        //mappedProducts.ReferenceAction = "Index";
+    //    //var mappedProducts = new PagedViewModel<ProductResponse>
+    //    //{
+    //    //    List = response.Data,
+    //    //    PageIndex = request.PageNumber,
+    //    //    PageSize = request.PageSize,
+    //    //    Query = request.Query,
+    //    //    TotalResults = response.TotalCount
+    //    //};
+    //    //ViewBag.Search = q;
+    //    //mappedProducts.ReferenceAction = "Index";
 
-        //return View(mappedProducts);
+    //    //return View(mappedProducts);
 
-        return View();
-    }
+    //    return View();
+    //}
 
-    public async Task<IActionResult> Details(Guid id)
-    {
-        var response = await GetById(id);
-        if (response is null)
-        {
-            return NotFound();
-        }
-        return View(response);
-    }
+    //public async Task<IActionResult> Details(Guid id)
+    //{
+    //    var response = await GetById(id);
+    //    if (response is null)
+    //    {
+    //        return NotFound();
+    //    }
+    //    return View(response);
+    //}
 
-    public IActionResult Create()
-    {
-        return View();
-    }
+    //public IActionResult Create()
+    //{
+    //    return View();
+    //}
 
-    [HttpPost]
-    [ValidateAntiForgeryToken]
-    public async Task<IActionResult> Create(CreateProductRequest request)
-    {
-        if (!ModelState.IsValid)
-        {
-            return View(request);
-        }
-        request.CompanyId = GetTenant();
-        //var result = await _productHandler.CreateAsync(request);
+    //[HttpPost]
+    //[ValidateAntiForgeryToken]
+    //public async Task<IActionResult> Create(CreateProductRequest request)
+    //{
+    //    if (!ModelState.IsValid)
+    //    {
+    //        return View(request);
+    //    }
+    //    request.CompanyId = GetTenant();
+    //    //var result = await _productHandler.CreateAsync(request);
 
-        //if (result != null && !result.IsSuccess)
-        //{
-        //    Notify(result.Message);
-        //    TempData["Failure"] = "Falha ao adicionar produto: " + string.Join("; ", await GetNotificationErrors());
-        //    return View(request);
-        //}
-        //if (!IsValidOperation())
-        //{
-        //    TempData["Failure"] = "Falha ao adicionar produto: " + string.Join("; ", await GetNotificationErrors());
-        //    return View(request);
-        //}
-        TempData["Success"] = "Produto adicionado com sucesso!";
-        return RedirectToAction(nameof(Index));
-    }
+    //    //if (result != null && !result.IsSuccess)
+    //    //{
+    //    //    Notify(result.Message);
+    //    //    TempData["Failure"] = "Falha ao adicionar produto: " + string.Join("; ", await GetNotificationErrors());
+    //    //    return View(request);
+    //    //}
+    //    //if (!IsValidOperation())
+    //    //{
+    //    //    TempData["Failure"] = "Falha ao adicionar produto: " + string.Join("; ", await GetNotificationErrors());
+    //    //    return View(request);
+    //    //}
+    //    TempData["Success"] = "Produto adicionado com sucesso!";
+    //    return RedirectToAction(nameof(Index));
+    //}
 
-    public async Task<IActionResult> Edit(Guid id)
-    {
-        var response = await GetById(id);
-        if (response is null)
-        {
-            return NotFound();
-        }
-        //var request = new UpdateProductRequest(id, response.Data.CompanyId, response.Data.Title, response.Data.Description, response.Data.UnitaryValue, response.Data.IsActive);
+    //public async Task<IActionResult> Edit(Guid id)
+    //{
+    //    var response = await GetById(id);
+    //    if (response is null)
+    //    {
+    //        return NotFound();
+    //    }
+    //    //var request = new UpdateProductRequest(id, response.Data.CompanyId, response.Data.Title, response.Data.Description, response.Data.UnitaryValue, response.Data.IsActive);
 
-        //return View(request);
+    //    //return View(request);
 
-        return View();
-    }
+    //    return View();
+    //}
 
-    [HttpPost]
-    [ValidateAntiForgeryToken]
-    public async Task<IActionResult> Edit(Guid id, UpdateProductRequest request)
-    {
-        if (id != request.Id)
-        {
-            return NotFound();
-        }
-        if (!ModelState.IsValid)
-        {
-            return View(request);
-        }
+    //[HttpPost]
+    //[ValidateAntiForgeryToken]
+    //public async Task<IActionResult> Edit(Guid id, UpdateProductRequest request)
+    //{
+    //    if (id != request.Id)
+    //    {
+    //        return NotFound();
+    //    }
+    //    if (!ModelState.IsValid)
+    //    {
+    //        return View(request);
+    //    }
 
-        request.CompanyId = GetTenant();
-        //var result = await _productHandler.UpdateAsync(request);
+    //    request.CompanyId = GetTenant();
+    //    //var result = await _productHandler.UpdateAsync(request);
 
-        //if (result != null && !result.IsSuccess)
-        //{
-        //    Notify(result.Message);
-        //    TempData["Failure"] = "Falha ao atualizar produto: " + string.Join("; ", await GetNotificationErrors());
-        //    return View(request);
-        //}
-        //if (!IsValidOperation())
-        //{
-        //    TempData["Failure"] = "Falha ao atualizar produto: " + string.Join("; ", await GetNotificationErrors());
-        //    return View(request);
-        //}
-        TempData["Success"] = "Produto atualizado com sucesso!";
-        return RedirectToAction(nameof(Index));
-    }
+    //    //if (result != null && !result.IsSuccess)
+    //    //{
+    //    //    Notify(result.Message);
+    //    //    TempData["Failure"] = "Falha ao atualizar produto: " + string.Join("; ", await GetNotificationErrors());
+    //    //    return View(request);
+    //    //}
+    //    //if (!IsValidOperation())
+    //    //{
+    //    //    TempData["Failure"] = "Falha ao atualizar produto: " + string.Join("; ", await GetNotificationErrors());
+    //    //    return View(request);
+    //    //}
+    //    TempData["Success"] = "Produto atualizado com sucesso!";
+    //    return RedirectToAction(nameof(Index));
+    //}
 
-    public async Task<IActionResult> Delete(Guid id)
-    {
-        var response = await GetById(id);
-        if (response is null)
-        {
-            return NotFound();
-        }
+    //public async Task<IActionResult> Delete(Guid id)
+    //{
+    //    var response = await GetById(id);
+    //    if (response is null)
+    //    {
+    //        return NotFound();
+    //    }
 
-        return View(response);
-    }
+    //    return View(response);
+    //}
 
-    [HttpPost, ActionName("Delete")]
-    [ValidateAntiForgeryToken]
-    public async Task<IActionResult> DeleteConfirmed(Guid id)
-    {
-        var response = await GetById(id);
-        if (response is null)
-        {
-            return NotFound();
-        }
+    //[HttpPost, ActionName("Delete")]
+    //[ValidateAntiForgeryToken]
+    //public async Task<IActionResult> DeleteConfirmed(Guid id)
+    //{
+    //    var response = await GetById(id);
+    //    if (response is null)
+    //    {
+    //        return NotFound();
+    //    }
 
-        await _productHandler.DeleteAsync(new DeleteProductRequest { Id = id });
+    //    await _productHandler.DeleteAsync(new DeleteProductRequest { Id = id });
 
-        if (!IsValidOperation())
-        {
-            TempData["Failure"] = "Falha ao deletar produto: " + string.Join("; ", await GetNotificationErrors());
-            return View(response);
-        }
+    //    if (!IsValidOperation())
+    //    {
+    //        TempData["Failure"] = "Falha ao deletar produto: " + string.Join("; ", await GetNotificationErrors());
+    //        return View(response);
+    //    }
 
-        TempData["Success"] = "Produto deletado com sucesso!";
-        return RedirectToAction(nameof(Index));
-    }
+    //    TempData["Success"] = "Produto deletado com sucesso!";
+    //    return RedirectToAction(nameof(Index));
+    //}
 
-    private Guid GetTenant()
-    {
-        return _appUser.GetTenantId();
-    }
-    private async Task<ProductResponse> GetById(Guid id)
-    {
-        return await _productHandler.GetByIdAsync(new GetProductByIdRequest { Id = id });
-    }
+    //private Guid GetTenant()
+    //{
+    //    return _appUser.GetTenantId();
+    //}
+    //private async Task<ProductResponse> GetById(Guid id)
+    //{
+    //    return await _productHandler.GetByIdAsync(new GetProductByIdRequest { Id = id });
+    //}
 }
