@@ -1,108 +1,8 @@
-﻿using EMS.WebApp.MVC.Extensions;
-using System.ComponentModel;
+﻿using System.ComponentModel;
 using System.ComponentModel.DataAnnotations;
 
 namespace EMS.WebApp.MVC.Models;
 
-public class UserViewModel
-{
-    public Guid Id { get; set; }
-    public Guid CompanyId { get; set; }
-    [Required(ErrorMessage = "O campo {0} é obrigatório")]
-    [DisplayName("Nome")]
-    public string Name { get; set; }
-
-    [Required(ErrorMessage = "O campo {0} é obrigatório")]
-    [DisplayName("Sobrenome")]
-    public string LastName { get; set; }
-
-    [Required(ErrorMessage = "O campo {0} é obrigatório")]
-    [EmailAddress(ErrorMessage = "O campo {0} está em formato inválido")]
-    [DisplayName("E-mail")]
-    public string Email { get; set; }
-
-    [Required(ErrorMessage = "O campo {0} é obrigatório")]
-    [DisplayName("Celular")]
-    public string PhoneNumber { get; set; }
-
-    [Cpf]
-    [Required(ErrorMessage = "O campo {0} é obrigatório")]
-    [DisplayName("CPF")]
-    public string Cpf { get; set; }
-
-    [DisplayName("Tipo")]
-    public string Role { get; set; }
-
-    [DisplayName("Ativo")]
-    public bool IsActive { get; set; }
-
-    [DisplayName("Data de Cadastro")]
-    public DateTime CreatedAt { get; set; }
-
-    [DisplayName("Data de Modificação")]
-    public DateTime UpdatedAt { get; set; }
-
-
-    public UserViewModel() { }
-
-    public UserViewModel(Guid id, Guid companyId, string name, string lastName, string email, string phoneNumber, string cpf, string role)
-    {
-        Id = id;
-        CompanyId = companyId;
-        Name = name;
-        LastName = lastName;
-        Email = email;
-        PhoneNumber = phoneNumber;
-        Cpf = cpf;
-        IsActive = true;
-        Role = role;
-    }
-    public UserViewModel(Guid id, Guid companyId, string name, string lastName, string email, string phoneNumber, string cpf, string role, DateTime createdAt, DateTime updatedAt)
-    {
-        Id = id;
-        CompanyId = companyId;
-        Name = name;
-        LastName = lastName;
-        Email = email;
-        PhoneNumber = phoneNumber;
-        Cpf = cpf;
-        IsActive = true;
-        Role = role;
-        CreatedAt = createdAt;
-        UpdatedAt = updatedAt;
-    }
-}
-
-public class RegisterUserViewModel
-{
-    [Required(ErrorMessage = "O campo {0} é obrigatório")]
-    public Guid Id { get; set; }
-
-    [DisplayName("Cargo")]
-    public string Role { get; set; } = string.Empty;
-
-    [Required(ErrorMessage = "O campo {0} é obrigatório")]
-    [EmailAddress(ErrorMessage = "O campo {0} está em formato inválido")]
-    [DisplayName("E-mail")]
-    public string Email { get; set; } = string.Empty;
-
-    [Required(ErrorMessage = "O campo {0} é obrigatório")]
-    [StringLength(100, ErrorMessage = "O campo {0} precisa ter entre {2} e {1} caracteres", MinimumLength = 6)]
-    [DisplayName("Senha")]
-    public string Password { get; set; } = string.Empty;
-}
-
-public class LoginUserViewModel
-{
-    [Required(ErrorMessage = "O campo {0} é obrigatório.")]
-    [EmailAddress(ErrorMessage = "O campo {0} está em formato inválido.")]
-    public string Email { get; set; } = string.Empty;
-
-    [Required(ErrorMessage = "O campo {0} é obrigatório.")]
-    [StringLength(100, ErrorMessage = "O campo {0} precisa ter entre {2} e {1} caracteres.", MinimumLength = 6)]
-    [DisplayName("Senha")]
-    public string Password { get; set; } = string.Empty;
-}
 
 public class UpdateUserViewModel
 {
