@@ -1,23 +1,17 @@
-using EMS.WebApp.Business.Configuration;
-using EMS.WebApp.Data.Configuration;
-using EMS.WebApp.Identity.Configuration;
-using EMS.WebApp.MVC.Configuration;
 using EMS.Core.Configuration;
+using EMS.WebApp.MVC.Configuration;
 
 var builder = WebApplication.CreateBuilder(args);
 
 builder.Services.AddMvcConfiguration(builder.Configuration);
 
 builder.Services.RegisterMVCServices()
-                .RegisterCoreServices()
-                .RegisterDataServices(builder.Configuration)
-                .RegisterIdentityServices(builder.Configuration)
-                .RegisterBusinessServices();
+                .RegisterCoreServices();
+
+builder.Services.AddIdentityConfiguration();
 
 var app = builder.Build();
 
 app.UseMvcConfiguration(app.Environment, app);
-
-app.CheckAndApplyDatabaseMigrations(app.Services);
 
 app.Run();
