@@ -1,4 +1,5 @@
-﻿@model LoginUserViewModel
+﻿@using EMS.Core.Requests.Identities
+@model LoginUserRequest
 
 @{
     ViewData["Title"] = "Log in";
