@@ -20,7 +20,7 @@
         <!-- step zero -->
         <div class="step">
             <h3 class="text-center mb-4">Confirmar Plano</h3>
-            <input type="hidden" asp-for="@Model.RegisterCompany.Company.PlanId" value="@Model.Plan.Id" />
+            <input type="hidden" asp-for="@Model.CreateCompanyAndUserRequest.Company.PlanId" value="@Model.Plan.Id" />
             <div class="card-body text-center">
                 <h4 class="title">@Model.Plan.Title</h4>
                 <p class="card-text">@Model.Plan.Subtitle</p>
@@ -44,12 +44,12 @@
         <div class="step">
             <h3 class="text-center mb-4">Dados da Empresa</h3>
             <div class="form-group">
-                <input asp-for="@Model.RegisterCompany.Company.Name" placeholder="Nome da Empresa" class="form-control" />
-                <span asp-validation-for="@Model.RegisterCompany.Company.Name" class="text-danger"></span>
+                <input asp-for="@Model.CreateCompanyAndUserRequest.Company.Name" placeholder="Nome da Empresa" class="form-control" />
+                <span asp-validation-for="@Model.CreateCompanyAndUserRequest.Company.Name" class="text-danger"></span>
             </div>
             <div class="form-group">
-                <input asp-for="@Model.RegisterCompany.Company.Document" placeholder="CPF/CNPJ" class="form-control" />
-                <span asp-validation-for="@Model.RegisterCompany.Company.Document" class="text-danger"></span>
+                <input asp-for="@Model.CreateCompanyAndUserRequest.Company.Document" placeholder="CPF/CNPJ" class="form-control" />
+                <span asp-validation-for="@Model.CreateCompanyAndUserRequest.Company.Document" class="text-danger"></span>
             </div>
         </div>
 
@@ -57,20 +57,20 @@
         <div class="step">
             <h3 class="text-center mb-4">Dados Pessoais</h3>
             <div class="form-group">
-                <input asp-for="@Model.RegisterCompany.Employee.Name" placeholder="Nome" class="form-control" />
-                <span asp-validation-for="@Model.RegisterCompany.Employee.Name" class="text-danger"></span>
+                <input asp-for="@Model.CreateCompanyAndUserRequest.Employee.Name" placeholder="Nome" class="form-control" />
+                <span asp-validation-for="@Model.CreateCompanyAndUserRequest.Employee.Name" class="text-danger"></span>
             </div>
             <div class="form-group">
-                <input asp-for="@Model.RegisterCompany.Employee.LastName" placeholder="Sobrenome" class="form-control" />
-                <span asp-validation-for="@Model.RegisterCompany.Employee.LastName" class="text-danger"></span>
+                <input asp-for="@Model.CreateCompanyAndUserRequest.Employee.LastName" placeholder="Sobrenome" class="form-control" />
+                <span asp-validation-for="@Model.CreateCompanyAndUserRequest.Employee.LastName" class="text-danger"></span>
             </div>
             <div class="form-group">
-                <input asp-for="@Model.RegisterCompany.Employee.Document" placeholder="CPF" class="form-control" />
-                <span asp-validation-for="@Model.RegisterCompany.Employee.Document" class="text-danger"></span>
+                <input asp-for="@Model.CreateCompanyAndUserRequest.Employee.Document" placeholder="CPF" class="form-control" />
+                <span asp-validation-for="@Model.CreateCompanyAndUserRequest.Employee.Document" class="text-danger"></span>
             </div>
             <div class="form-group">
-                <input asp-for="@Model.RegisterCompany.Employee.PhoneNumber" placeholder="Celular" class="form-control" />
-                <span asp-validation-for="@Model.RegisterCompany.Employee.PhoneNumber" class="text-danger"></span>
+                <input asp-for="@Model.CreateCompanyAndUserRequest.Employee.PhoneNumber" placeholder="Celular" class="form-control" />
+                <span asp-validation-for="@Model.CreateCompanyAndUserRequest.Employee.PhoneNumber" class="text-danger"></span>
             </div>
         </div>
 
@@ -78,18 +78,18 @@
         <div class="step">
             <h3 class="text-center mb-4">Configurações da Conta</h3>
             <div class="form-group">
-                <input asp-for="@Model.RegisterCompany.Employee.Email" placeholder="E-mail" class="form-control" />
-                <span asp-validation-for="@Model.RegisterCompany.Employee.Email" class="text-danger"></span>
+                <input asp-for="@Model.CreateCompanyAndUserRequest.User.Email" placeholder="E-mail" class="form-control" />
+                <span asp-validation-for="@Model.CreateCompanyAndUserRequest.User.Email" class="text-danger"></span>
             </div>
 
             <div class="form-group">
-                <input type="password" asp-for="@Model.RegisterCompany.Password" placeholder="Senha" class="form-control" />
-                <span asp-validation-for="@Model.RegisterCompany.Password" class="text-danger"></span>
+                <input type="password" asp-for="@Model.CreateCompanyAndUserRequest.User.Password" placeholder="Senha" class="form-control" />
+                <span asp-validation-for="@Model.CreateCompanyAndUserRequest.User.Password" class="text-danger"></span>
             </div>
 
             <div class="form-group">
-                <input type="password" asp-for="@Model.RegisterCompany.ConfirmPassword" placeholder="Confirme sua senha" class="form-control" />
-                <span asp-validation-for="@Model.RegisterCompany.ConfirmPassword" class="text-danger"></span>
+                <input type="password" asp-for="@Model.CreateCompanyAndUserRequest.User.ConfirmPassword" placeholder="Confirme sua senha" class="form-control" />
+                <span asp-validation-for="@Model.CreateCompanyAndUserRequest.User.ConfirmPassword" class="text-danger"></span>
             </div>
         </div>
 
