@@ -1,6 +1,5 @@
-﻿@using EMS.Core.Responses;
-@using EMS.Core.Responses.Clients;
-@model Response<ClientResponse>
+﻿@using EMS.Core.Responses.Clients;
+@model ClientResponse
 
 @{
     Layout = "_LayoutAdmin";
@@ -10,7 +9,7 @@
 <div class="card border-0 col-md-12">
     <div class="card-header d-flex justify-content-between">
         <h5 class="card-title">
-            Detalhes do Cliente: @Model.Data.Name @Model.Data.LastName
+            Detalhes do Cliente: @Model.Name @Model.LastName
         </h5>
     </div>
     <div class="card-body">
@@ -18,45 +17,45 @@
             <div class="col-md-4">
                 <dl class="row">
                     <dt class="col-sm-6">
-                        @Html.DisplayNameFor(model => model.Data.Name)
+                        @Html.DisplayNameFor(model => model.Name)
                     </dt>
                     <dd class="col-sm-6">
-                        @Html.DisplayFor(model => model.Data.Name)
+                        @Html.DisplayFor(model => model.Name)
                     </dd>
                     <dt class="col-sm-6">
-                        @Html.DisplayNameFor(model => model.Data.LastName)
+                        @Html.DisplayNameFor(model => model.LastName)
                     </dt>
                     <dd class="col-sm-6">
-                        @Html.DisplayFor(model => model.Data.LastName)
+                        @Html.DisplayFor(model => model.LastName)
                     </dd>
                     <dt class="col-sm-6">
-                        @Html.DisplayNameFor(model => model.Data.PhoneNumber)
+                        @Html.DisplayNameFor(model => model.PhoneNumber)
                     </dt>
                     <dd class="col-sm-6">
-                        @Html.DisplayFor(model => model.Data.PhoneNumber)
+                        @Html.DisplayFor(model => model.PhoneNumber)
                     </dd>
                     <dt class="col-sm-6">
-                        @Html.DisplayNameFor(model => model.Data.IsActive)
+                        @Html.DisplayNameFor(model => model.IsActive)
                     </dt>
                     <dd class="col-sm-6">
-                        @Html.DisplayFor(model => model.Data.IsActive)
+                        @Html.DisplayFor(model => model.IsActive)
                     </dd>
                     <dt class="col-sm-6">
-                        @Html.DisplayNameFor(model => model.Data.CreatedAt)
+                        @Html.DisplayNameFor(model => model.CreatedAt)
                     </dt>
                     <dd class="col-sm-6">
-                        @Html.DisplayFor(model => model.Data.CreatedAt)
+                        @Html.DisplayFor(model => model.CreatedAt)
                     </dd>
                     <dt class="col-sm-6">
-                        @Html.DisplayNameFor(model => model.Data.UpdatedAt)
+                        @Html.DisplayNameFor(model => model.UpdatedAt)
                     </dt>
                     <dd class="col-sm-6">
-                        @Html.DisplayFor(model => model.Data.UpdatedAt)
+                        @Html.DisplayFor(model => model.UpdatedAt)
                     </dd>
                 </dl>
                 <div class="form-group mt-3">
                     <a asp-action="Index" class="btn btn-warning me-3">Voltar</a>
-                    <a asp-action="Edit" asp-route-id="@Model.Data.Id" class="btn btn-primary">Editar</a>
+                    <a asp-action="Edit" asp-route-id="@Model.Id" class="btn btn-primary">Editar</a>
                 </div>
             </div>
         </div>
