@@ -1,4 +1,4 @@
-﻿@model EMS.WebApp.Business.Models.Company
+﻿@* @model EMS.WebApp.Business.Models.Company
 
 @{
     ViewData["Title"] = "Create";
@@ -50,3 +50,4 @@
 @section Scripts {
     @{await Html.RenderPartialAsync("_ValidationScriptsPartial");}
 }
+ *@
