@@ -1,4 +1,4 @@
-﻿@model EMS.WebApp.Business.Models.Company
+﻿@* @model EMS.WebApp.Business.Models.Company
 
 @{
     ViewData["Title"] = "Delete";
@@ -49,3 +49,4 @@
         <a asp-action="Index">Back to List</a>
     </form>
 </div>
+ *@
