@@ -1,4 +1,4 @@
-﻿@model EMS.WebApp.Business.Models.Company
+﻿@* @model EMS.WebApp.Business.Models.Company
 
 @{
     ViewData["Title"] = "Details";
@@ -46,3 +46,4 @@
     <a asp-action="Edit" asp-route-id="@Model.Id">Edit</a> |
     <a asp-action="Index">Back to List</a>
 </div>
+ *@
