@@ -1,4 +1,4 @@
-﻿@model EMS.WebApp.Business.Models.Company
+﻿@* @model EMS.WebApp.Business.Models.Company
 
 @{
     ViewData["Title"] = "Edit";
@@ -52,3 +52,4 @@
 @section Scripts {
     @{await Html.RenderPartialAsync("_ValidationScriptsPartial");}
 }
+ *@
