@@ -1,4 +1,4 @@
-﻿@model IEnumerable<EMS.WebApp.Business.Models.Company>
+﻿@* @model IEnumerable<EMS.WebApp.Business.Models.Company>
 
 @{
     ViewData["Title"] = "Index";
@@ -57,3 +57,4 @@
 }
     </tbody>
 </table>
+ *@
