@@ -1,4 +1,4 @@
-﻿@using EMS.WebApp.Business.Utils;
+﻿@using EMS.Core.User
 @inject IAspNetUser AppUser
 @{
     Layout = "_LayoutAdmin";
