@@ -1,4 +1,4 @@
-﻿@using EMS.WebApp.Business.Utils;
+﻿@using EMS.Core.User
 @using EMS.WebApp.MVC.Extensions;
 @inject IAspNetUser AppUser
 
